Move Drive into daemon.dbus
