+++ pepper2/common/drive_types/no_action.py
@@ -8,7 +8,7 @@ from .drive_type import DriveType
 
 if TYPE_CHECKING:
     from pepper2.daemon.dbus.controller import Controller
-    from pepper2.common.drives import Drive
+    from pepper2.daemon.dbus.drive import Drive
 
 
 class NoActionDriveType(DriveType):
