+++ pepper2/common/drive_types/usercode.py
@@ -20,7 +20,7 @@ from .no_action import NoActionDriveType
 
 if TYPE_CHECKING:
     from pepper2.daemon.dbus.controller import Controller
-    from pepper2.common.drives import Drive
+    from pepper2.daemon.dbus.drive import Drive
 
 LOGGER = logging.getLogger(__name__)
