+++ pepper2/daemon/dbus/controller.py
@@ -11,7 +11,7 @@ from pydbus.generic import signal
 from pepper2 import __version__
 from pepper2.common.daemon_status import DaemonStatus
 from pepper2.common.drive_types import DRIVE_TYPES
-from pepper2.common.drives import DriveGroup
+from pepper2.daemon.dbus.drive import DriveGroup
 from pepper2.daemon.usercode_driver import CodeStatus, UserCodeDriver
 
 LOGGER = logging.getLogger(__name__)
