+++ tests/common/test_drives.py
@@ -2,7 +2,7 @@
 from pathlib import Path
 
 from pepper2.common.drive_types import NoActionDriveType
-from pepper2.common.drives import Drive
+from pepper2.daemon.dbus.drive import Drive
 
 
 def test_drive_instantiation() -> None:
